requirements.txt
numpy==1.24.3
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
usearch==2.9.2
openai==1.3.9
